feat: be able to use remote server
