+++ admin/src/main/ts/.proxyRemoteConfig.template.js
@@ -1,0 +1,14 @@
+/**
+ * This file is used to store the credentials for the application.
+ * You can use it to store the target URL and the cookies to use for the proxy if you want to use a remote server.
+ * @target: the URL of the remote server
+ * @xsrfToken: the XSRF token to use for the proxy
+ * @oneSessionId: the oneSessionId to use for the proxy
+ */
+
+module.exports = {
+  target: "https://recette-release.opendigitaleducation.com",
+  xsrfToken: "cead076f-fa7c-4d00-b14a-15060eaa79c2",
+  oneSessionId:
+    "3ec16e80-7eee-4371-8865-e99d795266bb:TEaAu2cwbd9kWimV9wNDBrI+bGk=",
+};
